fix: split select employees page to smaller components
